agrupadas constantes en archivo de constantes
